修正经济合同高级搜索的分公司、车队下拉列表为空的错误

+++ bc-business-spider/src/test/java/cn/bc/business/spider/Spider4GzTaxiXhDriverInfoTest.java
@@ -21,7 +21,7 @@ public class Spider4GzTaxiXhDriverInfoTest {
 
 		String value = "257506";
 		Spider<GzTaxiXhDriverInfo> spider = new Spider4GzTaxiXhDriverInfo(type,
-				value, userName, userPassword, rootPath,0);
+				value, userName, userPassword, rootPath, (int) 0);
 		GzTaxiXhDriverInfo info = spider.excute();
 		Assert.assertNotNull(info);
